refactor: NavigationBar에 search버튼 추가 및 pages 폴더 구조 변경
