+++ src/Router.tsx
@@ -1,14 +1,19 @@
 import { createBrowserRouter } from 'react-router-dom';
 
-import AccountPage from './pages/AccountPage';
+import AccountPage from './pages/account';
 import HomePage from './pages/HomePage';
-import MessagePage from './pages/MessagePage';
-import NotificationPage from './pages/NotificationPage';
+import MessagePage from './pages/message';
+import NotificationPage from './pages/notification';
+import SearchPage from './pages/search';
 
 export const router = createBrowserRouter([
   {
     path: '/',
     element: <HomePage />,
+  },
+  {
+    path: '/search',
+    element: <SearchPage />,
   },
   {
     path: '/message',
