+++ src/pages/search/index.tsx
@@ -1,0 +1,3 @@
+export default function SearchPage() {
+  return <section>Search</section>;
+}
